+++ ft_printf.h
@@ -7,6 +7,6 @@
 int	ft_printf(const char *str, ...);	
 int ft_print_char(int c);
 int	ft_print_string(const char *src);
-int	ft_print_hex(unsigned long num);
+int	ft_print_hex(unsigned int num);
 
 #endif
